+++ test/webpack-test/app.vue
@@ -1,0 +1,11 @@
+<template>
+  <hello-world></hello-world>
+</template>
+<script>
+import HelloWorld from './hello-world.vue';
+export default {
+  components: {
+    HelloWorld
+  }
+}
+</script>