+++ test/webpack-test/flags.js
@@ -1,0 +1,4 @@
+module.exports = {
+  A: true,
+  B: false
+}