+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## vue-flags-webpack-plugin changelog
+### v1.0.0
+* support modify flags in development mode(flags could be file path)
+* changed the way to ignore files(now you have to set "?flag")
+* namespace must be specified now
+* allow to use this plugin without vue
+* fix some bugs
 
 ### v0.1.2
 * remove `htmlparser2` and use vue `htmlparser`